backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shital-backend"
version = "0.1.0"
description = "Shital Hindu Temple ERP - Python Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "orjson>=3.9.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.31",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "anthropic>=0.28.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "structlog>=24.2.0",
    "sendgrid>=6.11.0",
    "stripe>=9.9.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    "Pillow>=10.3.0",
    "jinja2>=3.1.4",
    "arrow>=1.3.0",
    "ulid-py>=1.1.0",
]

[project.optional-dependencies]
pdf = [
    "weasyprint>=62.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shital"]

[tool.ruff]
line-length = 160
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
